Episode 40: Finishing physics

+++ src/main/java/components/GridLines.java
@@ -15,22 +15,22 @@ public class GridLines extends Component {
         Vector2f projectionSize = camera.getProjectionSize();
 
         // back 1 unit to draw from the 0 offset on the screen coordinate (try to remote -1 to see more)
-        int firstX = ((int)(cameraPos.x / Settings.GRID_WIDTH) - 1) * Settings.GRID_WIDTH;
-        int firstY = ((int)(cameraPos.y / Settings.GRID_HEIGHT) - 1) * Settings.GRID_HEIGHT;
+        float firstX = ((int)(cameraPos.x / Settings.GRID_WIDTH) - 1) * Settings.GRID_WIDTH;
+        float firstY = ((int)(cameraPos.y / Settings.GRID_HEIGHT) - 1) * Settings.GRID_HEIGHT;
 
         // Draw 2 more lines: before the offset, after the max size (width and height on the screen)
         int numVtLines = (int)(projectionSize.x * camera.getZoom() / Settings.GRID_WIDTH) + 2;
         int numHzLines = (int)(projectionSize.y * camera.getZoom() / Settings.GRID_HEIGHT) + 2;
 
         // Because of 2 more lines drawn, add 2 more GRID_HEIGHT and GRID_WIDTH
-        int height = (int)(projectionSize.y * camera.getZoom()) + Settings.GRID_HEIGHT  * 2;
-        int width = (int)(projectionSize.x * camera.getZoom()) + Settings.GRID_WIDTH * 2;
+        float height = (int)(projectionSize.y * camera.getZoom()) + Settings.GRID_HEIGHT  * 2;
+        float width = (int)(projectionSize.x * camera.getZoom()) + Settings.GRID_WIDTH * 2;
 
         int maxLines = Math.max(numVtLines, numHzLines);
         Vector3f color = new Vector3f(0.2f, 0.2f, 0.2f);
         for (int i = 0; i < maxLines; i++) {
-            int x = firstX + (Settings.GRID_WIDTH * i); // max X of the vertical line
-            int y = firstY + (Settings.GRID_HEIGHT * i); // max y of the horizontal line
+            float x = firstX + (Settings.GRID_WIDTH * i); // max X of the vertical line
+            float y = firstY + (Settings.GRID_HEIGHT * i); // max y of the horizontal line
 
             if (i < numVtLines) {
                 DebugDraw.addLine2D(new Vector2f(x, firstY), new Vector2f(x, firstY + height), color);
